# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Schulanmeldeverfahren API
  description: API zur Verwaltung des schulischen Anmeldeverfahrens.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
        code:
          type: string

    User:
      type: object
      properties:
        user_id:
          type: string
        username:
          type: string
        group_name:
          type: string
        permissions:
          type: array
          items:
            type: string

paths:
  /api/connection/status:
    get:
      summary: Get database connection status
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  configured:
                    type: boolean
                  host:
                    type: string

  /api/connection/connect:
    post:
      summary: Connect to MariaDB
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [host, database, username, password]
              properties:
                host:
                  type: string
                port:
                  type: integer
                database:
                  type: string
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Connected successfully

  /api/auth/login:
    post:
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'